Fix spoiler submit
Remove validation on the text field to prevent the validation that is
built into browsers from preventing the form submission.

The form submit would usually trigger a function that copies over the
content from the fancy editor into the plain (but hidden) text field.
But since the form submission is prevented, that function never runs and
never makes the form valid even if content was added.

The downside of this change is that you can submit with an empty text
field but I would say that is not a real issue for users of this
service.